extra component for congratulation
