+++ src/Components/Congratulation/Congratulation.scss
@@ -1,0 +1,4 @@
+.congratulation {
+    font-size: 30px;
+    text-transform: uppercase;
+}
